pyproject.toml
[project]
name = "cloudeagle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
